--- components/Header/header.js
@@ -1,1 +1,0 @@
-//This will be the root file that will return all the pages
